Payment T4 and tests completed.

+++ ECS.MemberManager.Core.BusinessObjects/Payment/PaymentEC_ext.cs
@@ -9,7 +9,7 @@ namespace ECS.MemberManager.Core.BusinessObjects
             base.AddBusinessRules();
             
             BusinessRules.AddRule(new MaxLength(NotesProperty,255));
-            BusinessRules.AddRule(new MinValue<decimal>(AmountProperty,0.01M));
+            BusinessRules.AddRule(new MinValue<double>(AmountProperty,0.01));
             BusinessRules.AddRule(new Required(LastUpdatedByProperty));
             BusinessRules.AddRule(new MaxLength(LastUpdatedByProperty,255));
         }
